requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
chardet==5.1.0
Django==4.2.3
django-cors-headers==4.2.0
django-enumfield==3.1
django-filter==22.1
django-graphql-jwt==0.3.4
django-rut==1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
graphene==3.2.2
graphene-django==3.1.3
graphene-django-extras==1.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
lml==0.1.0
promise==2.3
psycopg2==2.9.6
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-xls==0.7.0
PyJWT==2.7.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
rut-chile==2.0.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
texttable==1.6.7
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
